Move existing sources and layers to useMapLayers and pass needed params
